# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "TES SWAGGER ULBI",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/mhrndiva",
            "email": "714220050@std.ulbi.ac.id"
        },
        "version": "1.0"
    },
    "host": "ws-ats-714220050-cc57ecdf5b73.herokuapp.com",
    "basePath": "/",
    "paths": {
        "/insert": {
            "post": {
                "description": "Input data presensi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Presensi"
                ],
                "summary": "Insert data presensi.",
                "parameters": [
                    {
                        "description": "Payload Body [RAW]",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ReqPresensi"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Presensi"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/presensi": {
            "get": {
                "description": "Mengambil semua data presensi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Presensi"
                ],
                "summary": "Get All Data Presensi.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Presensi"
                        }
                    }
                }
            }
        },
        "/presensi/{id}": {
            "get": {
                "description": "Ambil per ID data presensi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Presensi"
                ],
                "summary": "Get By ID Data Presensi.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Masukan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.Presensi"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.Mahasiswa": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "alamat": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "jurusan": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "npm": {
                    "type": "integer"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "controller.Matkul": {
            "type": "object",
            "properties": {
                "dosen": {
                    "type": "string"
                },
                "jadwal": {
                    "type": "string"
                },
                "namamatkul": {
                    "type": "string"
                },
                "sks": {
                    "type": "integer"
                }
            }
        },
        "controller.Presensi": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "123456789"
                },
                "biodata": {
                    "$ref": "#/definitions/controller.Mahasiswa"
                },
                "checkin": {
                    "type": "string"
                },
                "datetime": {
                    "type": "string"
                },
                "matkul": {
                    "$ref": "#/definitions/controller.Matkul"
                },
                "npm": {
                    "type": "integer",
                    "example": 714220050
                }
            }
        },
        "controller.ReqPresensi": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "123456789"
                },
                "biodata": {
                    "$ref": "#/definitions/controller.Mahasiswa"
                },
                "checkin": {
                    "type": "string"
                },
                "datetime": {
                    "type": "string"
                },
                "matkul": {
                    "$ref": "#/definitions/controller.Matkul"
                },
                "npm": {
                    "type": "integer",
                    "example": 714220050
                }
            }
        }
    }
}